Shifted speaker test code into audiobait.go
