+++ audiobait.go
@@ -19,7 +19,10 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 package managementinterface
 
 import (
+	"bufio"
 	"encoding/json"
+	"errors"
+	"fmt"
 	"io/ioutil"
 	"log"
 	"net/http"
@@ -32,7 +35,11 @@ import (
 	"github.com/TheCacophonyProject/audiobait/audiofilelibrary"
 	"github.com/TheCacophonyProject/audiobait/playlist"
 	goconfig "github.com/TheCacophonyProject/go-config"
+	"github.com/gobuffalo/packr"
+	"github.com/gorilla/mux"
 )
+
+var audioBox = packr.NewBox("./audio")
 
 // Return recent log entries from the audiobait process
 func getAudiobaitLogEntries() string {
@@ -292,4 +299,110 @@ func AudiobaitLogEntriesHandler(w http.ResponseWriter, r *http.Request) {
 
 	json.NewEncoder(w).Encode(response)
 
+}
+
+// AudiobaitSoundsHandlerGen is a wrapper for the AudiobaitSoundsHandler function.
+func AudiobaitSoundsHandlerGen(conf *goconfig.Config) func(http.ResponseWriter, *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
+		AudiobaitSoundsHandler(w, r, conf)
+	}
+}
+
+// AudiobaitSoundsHandler attempts to play a sound on connected speaker(s) at the volume set in the schedule.
+func AudiobaitSoundsHandler(w http.ResponseWriter, r *http.Request, conf *goconfig.Config) {
+
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	response := make(map[string]string)
+
+	// Extract sound file name
+	fileName := mux.Vars(r)["fileName"]
+	volume, _ := strconv.Atoi(mux.Vars(r)["volume"])
+
+	// Get audiobait directory
+	audio := goconfig.DefaultAudio()
+	if err := conf.Unmarshal(goconfig.AudioKey, &audio); err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		log.Printf("audio output failed: %v", err)
+		response["result"] = fmt.Sprintf("Error: %v.", err.Error())
+		json.NewEncoder(w).Encode(response)
+		return
+	}
+
+	// Play the sound.
+	if output, err := playAudioBaitSound(audio, fileName, volume); err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		log.Printf("audio output failed: %v", err)
+		response["result"] = fmt.Sprintf("Error: %v. Output:\n%s", err.Error(), string(output))
+	} else {
+		w.WriteHeader(http.StatusOK)
+		response["result"] = string(output)
+	}
+
+	// Encode data to be sent back to html.
+	json.NewEncoder(w).Encode(response)
+}
+
+// Play the specified sound on the speaker.
+func playAudioBaitSound(audio goconfig.Audio, fileName string, volume int) ([]byte, error) {
+
+	var cmd *exec.Cmd
+	var soundFile []byte
+
+	// Set volume
+	err := setVolume(volume, audio)
+	if err != nil {
+		return nil, fmt.Errorf("unable to set the volume: %v", err)
+	}
+
+	// Either we're playing the default test sound, or a sound from the schedule.
+	if fileName == "test.wav" {
+		soundFile = audioBox.Bytes("test.wav")
+		if soundFile == nil {
+			return nil, errors.New("unable to load test audio")
+		}
+		cmd = exec.Command("play", "-t", "wav", "--norm=-3", "-q", "-")
+	} else {
+		// It's a sound specified in the schedule. Load the file from the audiobait directory.
+		soundFile, err = ioutil.ReadFile(filepath.Join(audio.Dir, fileName))
+		if soundFile == nil || err != nil {
+			return nil, errors.New("unable to load audio bait sound file: " + fileName)
+		}
+		// Get the file type e.g. .wav, .mp3 etc.
+		fileType := filepath.Ext(fileName)
+		if fileType == "" {
+			fileType = "wav"
+		}
+		cmd = exec.Command("play", "-t", fileType, "--norm=-3", "-q", "-")
+	}
+
+	stdin, err := cmd.StdinPipe()
+	if err != nil {
+		return nil, fmt.Errorf("unable to play audio: %v", err)
+	}
+
+	go func() {
+		defer stdin.Close()
+		w := bufio.NewWriter(stdin)
+		if _, err := w.Write(soundFile); err != nil {
+			log.Printf("unable to pass audio: %v", err)
+		}
+	}()
+
+	return cmd.CombinedOutput()
+}
+
+// Set the volume on the sound card.
+func setVolume(volume int, audio goconfig.Audio) error {
+	cmd := exec.Command(
+		"amixer",
+		"-c", fmt.Sprint(audio.Card),
+		"sset",
+		audio.VolumeControl,
+		fmt.Sprintf("%d%%", volume*10),
+	)
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		return fmt.Errorf("volume set failed: %v\noutput:\n%s", err, out)
+	}
+	return nil
 }
